OGC Refactoring Part 8 - Move research query Factory service to package /ogc/services/researchQuery
