+++ server/src/main/java/de/ingrid/igeserver/ogc/services/researchQuery/OgcApiResearchQuery.kt
@@ -17,7 +17,7 @@
  * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
-package de.ingrid.igeserver.ogc
+package de.ingrid.igeserver.ogc.services.researchQuery
 
 import de.ingrid.igeserver.ClientException
 import de.ingrid.igeserver.model.BoolFilter
